pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-corpus-builder"
version = "0.1.0"
description = "Production-grade web crawling and preprocessing pipeline that converts raw websites into RAG-ready, searchable document datasets."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Somesh"},
]
keywords = ["rag", "web-crawling", "nlp", "dataset", "preprocessing", "embeddings"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "httpx>=0.27,<1",
    "beautifulsoup4>=4.12,<5",
    "lxml>=5.0,<6",
    "lxml-html-clean>=0.4,<1",
    "trafilatura>=1.12,<2",
    "tiktoken>=0.7,<1",
    "rich>=13.0,<14",
    "click>=8.1,<9",
    "pydantic>=2.5,<3",
    "pydantic-settings>=2.1,<3",
    "pyyaml>=6.0,<7",
    "xxhash>=3.4,<4",
    "langdetect>=1.0.9,<2",
    "tqdm>=4.66,<5",
    "aiofiles>=24.0,<25",
    "tenacity>=8.2,<10",
    "datasets>=2.18,<4",
    "pyarrow>=15.0",
    "robotexclusionrulesparser>=1.7,<2",
    "django>=5.0,<6",
    "djangorestframework>=3.15,<4",
    "django-cors-headers>=4.3,<5",
]

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=3.0,<4",
    "torch>=2.2,<3",
]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "ruff>=0.4,<1",
    "mypy>=1.9,<2",
]

[project.scripts]
rag-corpus = "rag_corpus_builder.cli:main"
rag-corpus-web = "web.manage:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
